requirements.txt
et_xmlfile==2.0.0
openpyxl==3.1.2
Pillow==10.1.0
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
PyPDF2==3.0.1
PySide6==6.6.0
PySide6-Addons==6.6.0
PySide6-Essentials==6.6.0
python-barcode==0.15.1
reportlab==4.0.7
shiboken6==6.6.0
